# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CategoryResponse:
    properties:
      category_type:
        allOf:
        - $ref: '#/definitions/dto.CategoryTypeResponse'
        description: Optional
      category_type_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      language_code:
        $ref: '#/definitions/enums.PageLanguage'
      name:
        type: string
      publish_status:
        $ref: '#/definitions/enums.PublishStatus'
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  dto.CategoryTypeResponse:
    properties:
      children_count:
        additionalProperties:
          type: integer
        description: 'e.g., {"th": 2, "en": 1}'
        type: object
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        description: OpenAPI บอก nullable
        type: string
      type_code:
        type: string
      updated_at:
        type: string
    type: object
  dto.CategoryTypeWithDetailsResponse:
    properties:
      categories:
        description: List of Category (Detail) responses for the specified language
        items:
          $ref: '#/definitions/dto.CategoryResponse'
        type: array
      created_at:
        type: string
      id:
        description: CategoryType ID
        type: string
      is_active:
        type: boolean
      name:
        type: string
      type_code:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateCategoryTypeRequest:
    properties:
      is_active:
        description: Default true ใน model
        type: boolean
      name:
        type: string
      type_code:
        type: string
    required:
    - type_code
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Failed to parse user data
        type: string
      message:
        example: Something went wrong
        type: string
    type: object
...
  dto.UpdateCategoryTypeRequest:
    properties:
      is_active:
        type: boolean
      name:
        type: string
    type: object
...
  enums.PageLanguage:
    enum:
    - th
    - en
    type: string
    x-enum-varnames:
    - PageLanguageTH
    - PageLanguageEN
...
  enums.PublishStatus:
    enum:
    - UnPublished
    - Published
    type: string
    x-enum-varnames:
    - PublishStatusNotPublished
    - PublishStatusPublished
...
info:
  contact: {}
  description: This API supports both Web and admin (CMS) operations, including authentication,
    page management.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /cms/category-types:
    get:
      description: Lists all category types, optionally filtered by isActive status.
        Includes count of categories per language.
      parameters:
      - description: Filter by active status
        in: query
        name: isActive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.CategoryTypeResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Category Types
      tags:
      - CMS - Category Types
    post:
      consumes:
      - application/json
      description: Creates a new category type
      parameters:
      - description: Category Type Data
        in: body
        name: category_type_data
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCategoryTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CategoryTypeResponse'
        "400":
          description: Validation Error or Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict - TypeCode already exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create Category Type
      tags:
      - CMS - Category Types
  /cms/category-types/{categoryTypeId}/categories:
    get:
      description: Retrieves all categories (details) associated with a specific category
        type, filtered by language.
      parameters:
      - description: Category Type ID (UUID)
        in: path
        name: categoryTypeId
        required: true
        type: string
      - description: Language code (th or en)
        enum:
        - th
        - en
        in: query
        name: lang
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryTypeWithDetailsResponse'
        "400":
          description: Invalid ID or language format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Category type not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List Categories for a Specific Type
      tags:
      - CMS - Category Types
  /cms/category-types/{id}:
    delete:
      description: Deletes a category type by its ID. Fails if categories are associated
        with it.
      parameters:
      - description: Category Type ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Category type not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict - Category type in use
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete Category Type
      tags:
      - CMS - Category Types
    get:
      description: Retrieves a specific category type by its ID, including count of
        categories per language.
      parameters:
      - description: Category Type ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryTypeResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Category type not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get Category Type by ID
      tags:
      - CMS - Category Types
    patch:
      consumes:
      - application/json
      description: Updates an existing category type by its ID.
      parameters:
      - description: Category Type ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Category Type Update Data
        in: body
        name: category_type_data
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CategoryTypeResponse'
        "400":
          description: Validation Error or Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Category type not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict - TypeCode already exists
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update Category Type
      tags:
      - CMS - Category Types
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"